Panamax becomes GoDutch!
The initial architecture will be on top of a single supervisor, to allow
them to communicate using pointers, instead of building one more IPC
mechanism. Therefore, Panamax became GoDutch, and a interface took
place, called "component", so GoDutch is able to handle a Container or a
Service. Both Container and NRPESrvc got changed to join the interface.